frontend : wip sidebar and navigation

+++ server/src/main/java/hupiat/intranet/server/core/errors/CommonControllerExceptions.java
@@ -4,7 +4,7 @@ import java.util.NoSuchElementException;
 
 public interface CommonControllerExceptions {
 
-    default void throwErrorNotFound(int id) {
+    default void throwErrorNotFound(long id) {
 	throw new NoSuchElementException("Could not found id : " + id);
     }
 }
